+++ app/Http/Controllers/Admin/ParameterController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Parameter;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Session;
 
 class ParameterController extends AdminController
@@ -19,9 +20,17 @@ class ParameterController extends AdminController
     {
         $input = $request->all();
         $input['value'] = $input['name'];
+        if (empty(trim($input['name']))) {
+            Session::flash('infomessage', 'У параметра должно быть не пустое название!!!');
+            return redirect('admin/parameter/group/'.$input['parameter_group_id']);
+        }
+        if (count(Parameter::where('name',$input['name'])->get())){
+            Session::flash('infomessage', 'Параметер с именем '.$input['name'].' уже существует!!!');
+            return redirect('admin/parameter/group/'.$input['parameter_group_id']);
+        }
 
         Parameter::create($input);
-        return redirect('admin/parameter/group/'.$input['parameter_group_id'] );
+        return redirect('admin/parameter/group/'.$input['parameter_group_id']);
     }
     public function edit($id)
     {
@@ -33,8 +42,18 @@ class ParameterController extends AdminController
     public function destroy ($id)
     {
         $parameter = Parameter::find($id);
+        $p = DB::table('article_parameter')
+            ->select('id')
+            ->where('parameter_id','=',$id)->get();
+
+        if(count($p)){
+            Session::flash('infomessage','Нельзя удалить параметер, у которго есть привязка к товарам');
+            return redirect('admin/parameter/group/'.$parameter['parameter_group_id']);
+        }
+
+
         if($parameter->delete())
-            Session::flash('infomessage',$parameter->name.' - deleted');
+            Session::flash('infomessage',$parameter->name.' - удален из базы');
 
         return redirect('admin/parameter/group/'.$parameter['parameter_group_id']);
     }
